# src/main/java/org/hu/reservation/controller/AdminController.java
package org.hu.reservation.controller;


import org.hu.reservation.dto.AvailableTablesResponse;
import org.hu.reservation.dto.Error;
import org.hu.reservation.entity.Reservation;
import org.hu.reservation.service.ReservationService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.text.SimpleDateFormat;
import java.time.LocalDate;
import java.time.ZoneId;
import java.util.Date;
import java.util.List;

@RestController
@RequestMapping("/api/admin")
@Validated
public class AdminController {

    private static final Logger logger = LoggerFactory.getLogger(AdminController.class);

    @Autowired
    private ReservationService reservationService;

    @GetMapping("/reservations")
    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
    public ResponseEntity<?> getReservations(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate reservationDate) {
        try {
            List<Reservation> reservations = reservationService.getReservationsByDate(reservationDate);
            return ResponseEntity.ok(reservations);
        } catch (Exception e) {
            logger.error(e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new Error("Server unavailable"));
        }
    }

    @GetMapping("/reservations/tables")
    @PreAuthorize("hasAuthority('ROLE_ADMIN')")
    // public ResponseEntity<?> getAvailableTables(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date) {
    public ResponseEntity<?> getAvailableTables(@RequestParam String date) {
        try {
            logger.info("Received date: " + date);
            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");

            Date date1 = sdf.parse(date);

            LocalDate localDate = date1.toInstant()
                    .atZone(ZoneId.systemDefault())
                    .toLocalDate();

            AvailableTablesResponse response = reservationService.getAvailableTables(localDate);
            logger.info("Response: " + response);
            return ResponseEntity.ok(response);
        } catch (Exception e) {
            logger.error(e.getMessage(), e);
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new Error("An error occurred while fetching available tables"));
        }
    }


}